+++ src/main/groovy/com/cedarsoftware/ncube/NCubeMutableClient.groovy
@@ -40,6 +40,8 @@ interface NCubeMutableClient extends NCubeClient
 
     Boolean assertPermissions(ApplicationID appId, String resource, Action action)
 
+    Map checkPermissions(ApplicationID appId, String resource, String[] actions)
+
     Boolean checkPermissions(ApplicationID appId, String resource, String action)
 
     Boolean isAppAdmin(ApplicationID appId)